Add prometheus default config

+++ apk/CONTROL/start-stop.sh
@@ -4,6 +4,11 @@
 . /usr/local/AppCentral/cappysan-prometheus/.env.install
 cd ${APKG_PKG_DIR:-/nonexistent} || exit 1
 
+function logger() {
+  echo "${@}" >&2
+  syslog --log 0 --level 0 --user SYSTEM --event "${@}"
+}
+
 export HOME=/share/Configuration/prometheus
 export PID_FILE=/var/run/prometheus.pid
 if test -f ${HOME}/env; then
@@ -24,8 +29,9 @@ export APKG_GROUP=${APKG_GROUP:-root}
 #   promtool check config /share/Configuration/prometheus/prometheus.yml
 ln -sf -T /usr/local/AppCentral/cappysan-prometheus/promtool /usr/local/bin/promtool
 
+# Create a chain key for prometheus
 cat /usr/builtin/etc/certificate/ssl.pem \
-    /usr/builtin/etc/certificate/ssl.crt   > /usr/local/AppCentral/cappysan-prometheus/chain.crt
+    /usr/builtin/etc/certificate/ssl.crt   > /usr/local/AppCentral/cappysan-prometheus/ssl.crt
 cp -f /usr/builtin/etc/certificate/ssl.key /usr/local/AppCentral/cappysan-prometheus/ssl.key
 chmod 600 /usr/local/AppCentral/cappysan-prometheus/ssl.key
 chown -R ${APKG_USER}:${APKG_GROUP} /usr/local/AppCentral/cappysan-prometheus/ssl.key
@@ -33,12 +39,9 @@ chown -R ${APKG_USER}:${APKG_GROUP} /usr/local/AppCentral/cappysan-prometheus/ss
 mkdir -p "${STORAGE_TSDB_PATH}"
 chown -R ${APKG_USER}:${APKG_GROUP} "${STORAGE_TSDB_PATH}"
 
-# Build a full chain
-cat /usr/builtin/etc/certificate/ssl.pem \
-    /usr/builtin/etc/certificate/ssl.crt > /usr/local/AppCentral/cappysan-prometheus/ssl.crt
-
 case $1 in
   start)
+    logger "[prometheus] Starting..."
     touch "${APKG_CFG_DIR}/active"
     start-stop-daemon -S -b -m -p ${PID_FILE} -c ${APKG_USER}:nogroup -x "./prometheus" -- \
        --config.file=${CONFIG_FILE} --storage.tsdb.path=${STORAGE_TSDB_PATH} \
@@ -48,10 +51,8 @@ case $1 in
     ;;
 
   stop)
-    if test -f "${APKG_CFG_DIR}/active"; then
-      rm -f "${APKG_CFG_DIR}/active"
-    fi
-
+    logger "[prometheus] Stopping..."
+    rm -f "${APKG_CFG_DIR}/active"
     if test -f ${PID_FILE}; then
       start-stop-daemon -K -p ${PID_FILE}
       rm -f ${PID_FILE}
@@ -59,11 +60,14 @@ case $1 in
     ;;
 
   reload)
-    if test -f ${PID_FILE}; then
-      touch "${APKG_CFG_DIR}/active"
-      kill -SIGHUP $(cat ${PID_FILE})
-    else
-      exit 1
+    if test -f "${APKG_CFG_DIR}/active"; then
+      if test -f ${PID_FILE}; then
+        logger "[prometheus] Reloading..."
+        touch "${APKG_CFG_DIR}/active"
+        kill -SIGHUP $(cat ${PID_FILE})
+      else
+        exit 1
+      fi
     fi
     ;;
